% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddt.R
\docType{data}
\name{ddt}
\alias{ddt}
\title{DDT data set}
\format{
A data frame with 144 rows and 5 columns:
\describe{ **EDIT FIELDS BELOW FOR VARIABLES**
  \item{DDT}{Amount of ddt}
  \item{MILE}{The mile}
  \item{RIVER}{River fish was caught in}
  \item{LENGTH}{Length of fish}
  \item{WEIGHT}{Weight of fish}
  \item{SPECIES}{Species of fish}
  ...
}
}
\source{
<https://www.amazon.com/Statistics-Engineering-Sciences-William...
}
\usage{
ddt
}
\description{
A complete data set of measurements of 144 fish caught on various rivers
}
\keyword{datasets}
