% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onewayTable.R
\name{onewayTable}
\alias{onewayTable}
\title{onewayTable}
\usage{
onewayTable(tab, alpha = 0.05)
}
\arguments{
\item{tab}{a one-way table of counts}

\item{alpha}{alpha level}
}
\value{
* named list with cell and combinations of cells probabilities
* barplots with Bonferonni intervals
}
\description{
A function which produces point and interval estimates for each cell probability (and combinations) and barplots with Bonferonni intervals
}
\examples{
\dontrun{onewayTable(tab = TABLE, alpha = 0.05)}
}
